Conflicts:
	packages/ifpack2/src/Ifpack2_AdditiveSchwarz_def.hpp

+++ packages/rythmos/doc/manual/generate_UnitTest_plots.py
@@ -69,8 +69,9 @@ def makefig(data, x_range=None, y_range=None, xy_graph=(0.9,0.98)):
   if x_range != None: lines.append('set xrange [%g:%g]\n' % x_range)
   if y_range != None: lines.append('set yrange [%g:%g]\n' % y_range)
   #lines.append('set key left top\n')
-  lines.append('set key at graph %g, graph %g Right\n' % (xy_graph[0], \
-                                                          xy_graph[1]) )
+  lines.append('set key right bottom\n')
+  #lines.append('set key at graph %g, graph %g Right\n' % (xy_graph[0], \
+  #                                                        xy_graph[1]) )
   first = True
   for datum in data:
     results_file, figure_name, order, time, error = datum
@@ -97,7 +98,8 @@ def makefig(data, x_range=None, y_range=None, xy_graph=(0.9,0.98)):
   for datum in data:
     results_file, figure_name, order, time, error = datum
     order_file = 'order%i.dat' %(order)
-    os.remove(order_file)
+    if os.path.exists(order_file):
+      os.remove(order_file)
 
 
 #===============================================================================
@@ -117,87 +119,107 @@ def main():
   opts, args = p.parse_args()
 
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ForwardEuler.dat')
   data = [(results, 'Forward Euler', 1.0, 0.1, 0.01)]
   makefig(data, (0.003,0.1), (1.0e-4,1.0e-1))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'BackwardEuler.dat')
   data = [(results, 'Backward Euler', 1.0, 0.1, 0.01)]
   makefig(data, (0.003,0.1), (1.0e-4,1.0e-1))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   data = []
   results = os.path.join(opts.directory, 'ImplicitBDF1.dat')
-  data.append((results, 'Implicit BDF1', 1.0, 0.025, 0.004))
+  data.append((results, 'IBDF1', 1.0, 0.025, 0.004))
   results = os.path.join(opts.directory, 'ImplicitBDF2.dat')
-  data.append((results, 'Implicit BDF2', 2.0, 0.025, 4.0e-5))
+  data.append((results, 'IBDF2', 2.0, 0.025, 4.0e-5))
   results = os.path.join(opts.directory, 'ImplicitBDF3.dat')
-  data.append((results, 'Implicit BDF3', 3.0, 0.025, 1.0e-6))
+  data.append((results, 'IBDF3', 3.0, 0.025, 1.0e-6))
   results = os.path.join(opts.directory, 'ImplicitBDF4.dat')
-  data.append((results, 'Implicit BDF4', 4.0, 0.025, 1.0e-8))
-  makefig(data, (0.0007,1.0), (1.0e-14,1.0e-1), (0.98,0.98))
+  data.append((results, 'IBDF4', 4.0, 0.025, 1.0e-8))
+  makefig(data, (0.0007,2.0), (1.0e-14,1.0e-1), (0.98,0.98))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_ForwardEuler.dat')
   data = [(results, 'ERK Forward Euler', 1.0, 0.1, 0.01)]
   makefig(data, (0.003,0.1), (1.0e-4,1.0e-1))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_4Stage.dat')
-  data = [(results, 'Explicit RK4', 4.0, 0.1, 1.0e-7)]
+  data = [(results, 'ERK4', 4.0, 0.1, 1.0e-7)]
   makefig(data, (0.003,0.1), (1.0e-14,1.0e-5))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_3_8_Rule.dat')
-  data = [(results, 'Explicit RK 3/8 Rule', 4.0, 0.1, 1.0e-7)]
+  data = [(results, 'ERK 3/8 Rule', 4.0, 0.1, 1.0e-7)]
   makefig(data, (0.003,0.1), (1.0e-14,1.0e-5))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_4Stage3OrderRunge.dat')
-  data = [(results, 'Explicit RK 4 Stage 3rd order by Runge', 3.0, 0.1, 1.0e-5)]
+  data = [(results, 'ERK4 3rd Order Runge', 3.0, 0.1, 1.0e-5)]
   makefig(data, (0.003,0.1), (1.0e-10,1.0e-4))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_3Stage3Order.dat')
-  data = [(results, 'Explicit RK 3 Stage 3rd order', 3.0, 0.1, 1.0e-5)]
+  data = [(results, 'RK3 3rd Order', 3.0, 0.1, 1.0e-5)]
   makefig(data, (0.003,0.1), (1.0e-10,1.0e-4))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_3Stage3OrderHeun.dat')
-  data = [(results, 'Explicit RK 3 Stage 3rd order by Heun', 3.0, 0.1, 1.0e-5)]
+  data = [(results, 'ERK3 3rd Order Heun', 3.0, 0.1, 1.0e-5)]
   makefig(data, (0.003,0.1), (1.0e-10,1.0e-4))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_3Stage3OrderTVD.dat')
-  data = [(results, 'Explicit RK 3 Stage 3rd order TVD', 3.0, 0.1, 1.0e-5)]
+  data = [(results, 'ERK3 3rd Order TVD', 3.0, 0.1, 1.0e-5)]
   makefig(data, (0.003,0.1), (1.0e-10,1.0e-4))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_2Stage2OrderRunge.dat')
-  data = [(results, 'Explicit RK 2 Stage 2nd order by Runge', 2.0, 0.1, 5.0e-4)]
+  data = [(results, 'ERK2 2nd Order Runge', 2.0, 0.1, 5.0e-4)]
   makefig(data, (0.003,0.1), (1.0e-7,1.0e-2))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'ERK_Trapezoidal.dat')
-  data = [(results, 'Explicit RK Trapezoidal', 2.0, 0.1, 5.0e-4)]
+  data = [(results, 'ERK Trapezoidal', 2.0, 0.1, 5.0e-4)]
   makefig(data, (0.003,0.1), (1.0e-7,1.0e-2))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'IRK_BackwardEuler.dat')
-  data = [(results, 'Implicit RK BackwardEuler', 1.0, 0.1, 0.01)]
+  data = [(results, 'IRK Backward Euler', 1.0, 0.1, 0.01)]
   makefig(data, (0.003,0.1), (1.0e-4,1.0e-1))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
+  data = []
+  results = os.path.join(opts.directory, 'IRK1StageTheta0.dat')
+  data.append((results, 'theta = 0', 1.0, 0.1, 0.01))
+  results = os.path.join(opts.directory, 'IRK1StageTheta.dat')
+  data.append((results, 'theta = 1/2', 2.0, 0.1, 5.0e-4))
+  results = os.path.join(opts.directory, 'IRK1StageTheta1.dat')
+  data.append((results, 'theta = 1', 1.0, 0.1, 0.01))
+  makefig(data, (0.003,0.1), (1.0e-8,1.0e-1))
+
+  # ---------------------------------------------------------------
+  data = []
+  results = os.path.join(opts.directory, 'IRK2StageTheta0.dat')
+  data.append((results, 'theta = 0+', 1.0, 0.1, 0.01))
+  results = os.path.join(opts.directory, 'IRK2StageTheta.dat')
+  data.append((results, 'theta = 1/2', 2.0, 0.1, 5.0e-4))
+  results = os.path.join(opts.directory, 'IRK2StageTheta1.dat')
+  data.append((results, 'theta = 1', 1.0, 0.1, 0.01))
+  makefig(data, (0.003,0.1), (1.0e-8,1.0e-1))
+
+  # ---------------------------------------------------------------
   data = []
   results = os.path.join(opts.directory, 'SDIRK_2Stage2Ordergp5.dat')
-  data.append((results, 'gamma = 0.5', 1.0, 0.1, 0.01))
+  data.append((results, 'g = 0.5', 1.0, 0.1, 0.01))
   results = os.path.join(opts.directory, 'SDIRK_2Stage2Order.dat')
-  data.append((results, 'gamma = (2-sqrt(2))/2', 2.0, 0.1, 5.0e-4))
+  data.append((results, 'g = (2-sqrt(2))/2', 2.0, 0.1, 5.0e-4))
   makefig(data, (0.003,0.1), (1.0e-7,1.0e-1))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   data = []
   results = os.path.join(opts.directory, 'SDIRK_2Stage3OrderAStable.dat')
   data.append((results, 'A-stable', 3.0, 0.1, 1.0e-5))
@@ -205,24 +227,24 @@ def main():
   data.append((results, 'L-stable', 2.0, 0.1, 5.0e-4))
   makefig(data, (0.003,0.1), (1.0e-10,1.0e-2))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'SDIRK_5Stage5Order.dat')
-  data = [(results, 'SDIRK 5 Stage 5th Order', 5.0, 0.1, 1.0e-09)]
+  data = [(results, 'SDIRK5 5th Order', 5.0, 0.1, 1.0e-09)]
   makefig(data, (0.003,0.1), (1.0e-15,1.0e-8))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'SDIRK_5Stage4Order.dat')
-  data = [(results, 'SDIRK 5 Stage 4th Order', 4.0, 0.1, 1.0e-07)]
+  data = [(results, 'SDIRK5 4th Order', 4.0, 0.1, 1.0e-07)]
   makefig(data, (0.003,0.1), (1.0e-14,1.0e-5))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'SDIRK_3Stage4Order.dat')
-  data = [(results, 'SDIRK 3 Stage 4th Order', 4.0, 0.1, 1.0e-07)]
+  data = [(results, 'SDIRK3 4th Order', 4.0, 0.1, 1.0e-07)]
   makefig(data, (0.003,0.1), (1.0e-14,1.0e-5))
 
-  # --------------------------------------------------------------- 
+  # ---------------------------------------------------------------
   results = os.path.join(opts.directory, 'DIRK_2Stage3Order.dat')
-  data = [(results, 'DIRK 2 Stage 3th Order', 3.0, 0.1, 1.0e-05)]
+  data = [(results, 'DIRK2 3rd Order', 3.0, 0.1, 1.0e-05)]
   makefig(data, (0.003,0.1), (1.0e-10,1.0e-4))
